Fix r6dumper build errors and compiler warnings
- Renamed `driver::interface` to `driver::g_interface` to avoid conflicts with the Windows `interface` macro.
- Updated `r6od.vcxproj` to use `$(DefaultPlatformToolset)` for better compatibility.
- Resolved multiple C4244, C4267, and C4477 warnings by using explicit casts and correct format specifiers.
- Fixed a sign-extension/truncation bug in `extract_relative_offset` by changing the return type to `uint64_t`.
- Resolved a shadow variable issue in `adder.h`.

+++ r6dumper/src/main.cpp
@@ -24,7 +24,7 @@ int main()
         return 1;
     }
 
-    base = driver::get_module_base(pid, L"RainbowSix.exe");
+    base = driver::get_module_base((uint64_t)pid, L"RainbowSix.exe");
 
     if (base == 0) {
         std::cout << "[-] failed to find module base!" << std::endl;
@@ -35,7 +35,7 @@ int main()
 
     memory = new uint8_t[seg_end(segment::data)];
 
-    if (driver::read_memory(handle, base, memory, seg_end(segment::data))) {
+    if (driver::read_memory(handle, base, memory, (uint32_t)seg_end(segment::data))) {
         std::cout << "error reading memory\n";
         std::cin.get();
         return 1;
